feat: moved ts components as svelte
ts components should be at src/components
svelte components were added
