+++ package.json
@@ -22,8 +22,10 @@
     "fix:prettier": "prettier -w ."
   },
   "dependencies": {
+    "@astrojs/node": "^9.3.0",
     "@astrojs/rss": "^4.0.12",
     "@astrojs/sitemap": "^3.4.1",
+    "@astrojs/svelte": "^7.1.0",
     "@astrolib/analytics": "^0.6.1",
     "@astrolib/seo": "^1.0.0-beta.8",
     "@fontsource-variable/inter": "^5.2.6",
@@ -62,6 +64,7 @@
     "mdast-util-to-string": "^4.0.0",
     "prettier": "^3.6.2",
     "prettier-plugin-astro": "^0.14.1",
+    "prettier-plugin-svelte": "^3.4.0",
     "reading-time": "^1.5.0",
     "sharp": "0.34.2",
     "tailwind-merge": "^3.3.1",
