+++ src/components/common/HydratedImage.svelte
@@ -1,0 +1,6 @@
+<script lang="ts">
+  export let src: string;
+  export let attributes: Record<string, any> = {};
+</script>
+
+<img {...attributes} {src} />
