+++ src/data/post/cogfunsearch-symbolic-cognition.mdx
@@ -1,5 +1,5 @@
 ---
-publishDate: 2025-03-30T00:00:01Z
+publishDate: 2025-03-30T09:00:01+09:00
 author: Younghoon Kim
 title: 'CogFunSearch: Symbolic Cognitive Models with LLMs'
 image: https://img-svr.nabiresearch.workers.dev/img/webp/ZqB593s.webp
