+++ src/data/post/neural-coding-olfaction.mdx
@@ -1,5 +1,5 @@
 ---
-publishDate: 2025-03-30T00:00:00Z
+publishDate: 2025-03-30T09:00:00+09:00
 author: Chaebeom Sheen
 title: A Neural Coding Model for Olfaction
 image: https://img-svr.nabiresearch.workers.dev/img/webp/cb2BnBy.webp
